Move assertUnreacheable to its own file, rename canFocus to shouldFocus
