+++ packages/teleterm/src/ui/utils/index.ts
@@ -2,3 +2,4 @@ export * from './getClusterName';
 export * from './uid';
 export * from './retryWithRelogin';
 export * from './getUserWithClusterName';
+export * from './assertUnreachable';
